requirements.txt
Django==2.1.7
djangorestframework==3.9.1
djangorestframework-simplejwt==4.3.0
Faker==1.0.6


